+++ src/main/java/com/avimedical/appointments/scheduling/adapters/database/ReasonCrudRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-interface ReasonCrudRepository extends CrudRepository<ReasonEntity, Long> {
+interface ReasonCrudRepository extends CrudRepository<ReasonEntity, Integer> {
 }
