+++ src/main/java/com/avimedical/appointments/scheduling/adapters/database/entities/ReasonEntity.java
@@ -21,7 +21,7 @@ public class ReasonEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id; // Should be UUID or other non-guessable key format in production
+    private Integer id; // Should be UUID or other non-guessable key format in production
     private String title;
     private String description;
 }
